[ARM] arm32ve: arm64: Fix cpu_vcpu_coproc for instruction emulation
The ARM instruction emulation library under arch/arm/cpu/common uses
cpu_vcpu_coproc provided by arm32, arm32ve, and arm64 to determing
coprocessor callbacks.

This patch fixes cpu_vcpu_coproc for arm32ve and arm64 because we now
have CP14 read/write functions available in both these ports.

+++ arch/arm/cpu/arm64/include/cpu_vcpu_sysregs.h
@@ -31,25 +31,25 @@
 bool cpu_vcpu_cp15_read(struct vmm_vcpu *vcpu,
 			arch_regs_t *regs,
 			u32 opc1, u32 opc2, u32 CRn, u32 CRm,
-			u64 *data);
+			u32 *data);
 
 /** Write one CP15 register */
 bool cpu_vcpu_cp15_write(struct vmm_vcpu *vcpu,
 			 arch_regs_t *regs,
 			 u32 opc1, u32 opc2, u32 CRn, u32 CRm,
-			 u64 data);
+			 u32 data);
 
 /** Read one CP14 register */
 bool cpu_vcpu_cp14_read(struct vmm_vcpu *vcpu,
 			arch_regs_t *regs,
 			u32 opc1, u32 opc2, u32 CRn, u32 CRm,
-			u64 *data);
+			u32 *data);
 
 /** Write one CP14 register */
 bool cpu_vcpu_cp14_write(struct vmm_vcpu *vcpu,
 			 arch_regs_t *regs,
 			 u32 opc1, u32 opc2, u32 CRn, u32 CRm,
-			 u64 data);
+			 u32 data);
 
 /** Read one system register */
 bool cpu_vcpu_sysregs_read(struct vmm_vcpu *vcpu,
